requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
langchain-community
pypdf
beautifulsoup4
arxiv
PyMuPDF
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_groq
uvicorn
langserve
streamlit